+++ src/main/scala/org/broadinstitute/hail/driver/ImputeGender.scala
@@ -16,6 +16,12 @@ object ImputeGender extends Command {
 
     @Args4jOption(required = false, name = "-e", aliases = Array("--excludepar"), usage = "Exclude Pseudoautosomal regions")
     var excludePAR: Boolean = false
+
+    @Args4jOption(required = false, name = "-x", aliases = Array("--femalethreshold"), usage = "Samples are called females if F < femaleThreshold (Default = 0.2)")
+    var fFemaleThreshold: Double = 0.2
+
+    @Args4jOption(required = false, name = "-y", aliases = Array("--malethreshold"), usage = "Samples are called males if F > maleThreshold (Default = 0.8)")
+    var fMaleThreshold: Double = 0.8
   }
 
   def newOptions = new Options
@@ -28,11 +34,12 @@ object ImputeGender extends Command {
 
   def run(state: State, options: Options): State = {
 
-    val result = ImputeGenderPlink.imputeGender(state.vds, options.mafThreshold, options.excludePAR).result.collect().toMap
+    val result = ImputeGenderPlink.imputeGender(state.vds, options.mafThreshold, options.excludePAR)
+      .result(options.fFemaleThreshold, options.fMaleThreshold).collect().toMap
 
     val signature = TStruct("F" -> TDouble, "E" -> TDouble, "O" -> TDouble, "N" -> TInt, "T" -> TInt, "imputedSex" -> TInt)
 
-    val (newSAS, insertSexCheck) = state.vds.saSignature.insert(signature, "imputesex")
+    val (newSAS, insertSexCheck) = state.vds.saSignature.insert(signature, "imputegender")
     val newSampleAnnotations = state.vds.sampleIdsAndAnnotations
       .map { case (s, sa) =>
         insertSexCheck(sa, result.get(s))